added builder coverage, moved test builder methods to test src
