+++ src/main/java/com/upserve/uppend/AppendOnlyStoreBuilder.java
@@ -74,31 +74,6 @@ public class AppendOnlyStoreBuilder extends FileStoreBuilder<AppendOnlyStoreBuil
         );
     }
 
-    public static AppendOnlyStoreBuilder getDefaultTestBuilder() {
-        return getDefaultTestBuilder(ForkJoinPool.commonPool());
-    }
-
-    public static AppendOnlyStoreBuilder getDefaultTestBuilder(ExecutorService testService) {
-        return new AppendOnlyStoreBuilder()
-                .withStoreName("test")
-                .withBlobPageSize(64 * 1024)
-                .withBlobsPerBlock(30)
-                .withInitialBlobCacheSize(64)
-                .withMaximumBlobCacheSize(128)
-                .withBlobCacheExecutorService(testService)
-                .withInitialLookupKeyCacheSize(64)
-                .withMaximumLookupKeyCacheWeight(100 * 1024)
-                .withLookupKeyCacheExecutorService(testService)
-                .withInitialMetaDataCacheSize(64)
-                .withMaximumMetaDataCacheWeight(100 * 1024)
-                .withLookupMetaDataCacheExecutorService(testService)
-                .withLongLookupHashSize(16)
-                .withLookupPageSize(16 * 1024)
-                .withLookupPageCacheExecutorService(testService)
-                .withCacheMetrics();
-
-    }
-
     public int getBlobsPerBlock() {
         return blobsPerBlock;
     }
