+++ src/test/java/com/upserve/uppend/AppendOnlyStorePartitionSizeTest.java
@@ -14,7 +14,7 @@ public class AppendOnlyStorePartitionSizeTest {
     private final Path path = Paths.get("build/test/file-append-only-partition-size");
 
     private AppendOnlyStore newStore() {
-        return AppendOnlyStoreBuilder.getDefaultTestBuilder().withDir(path.resolve("store-path")).withPartitionSize(2).build(false);
+        return TestHelper.getDefaultTestBuilder().withDir(path.resolve("store-path")).withPartitionSize(2).build(false);
     }
     private AppendOnlyStore store;
